chore: refactor to use registerEventSubscribers
