+++ src/Kernel/index.ts
@@ -1,2 +1,3 @@
 export * from './types';
+export { registerEventSubscribers } from '@/Kernel/register-event-subscribers';
 export { initializeRequestScopeStorage, useEmit, useRequest, useResponse } from '@/Http/Scope/request-scope-storage';
